`InteractionType` should be `Int` Saves casting

+++ src/commonMain/kotlin/interaction/InteractionType.kt
@@ -1,9 +1,9 @@
 package cloud.drakon.tempest.interaction
 
 object InteractionType {
-    const val PING: Byte = 1
-    const val APPLICATION_COMMAND: Byte = 2
-    const val MESSAGE_COMPONENT: Byte = 3
-    const val APPLICATION_COMMAND_AUTOCOMPLETE: Byte = 4
-    const val MODAL_SUBMIT: Byte = 5
+    const val PING: Int = 1
+    const val APPLICATION_COMMAND: Int = 2
+    const val MESSAGE_COMPONENT: Int = 3
+    const val APPLICATION_COMMAND_AUTOCOMPLETE: Int = 4
+    const val MODAL_SUBMIT: Int = 5
 }
